# oapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the NBRGLM Nexeres API Spec documentation.",
        "title": "NBRGLM Nexeres API Spec",
        "termsOfService": "https://nbrglm.com/nexeres/terms",
        "contact": {
            "name": "NBRGLM Support",
            "url": "https://nbrglm.com/support",
            "email": "contact@nbrglm.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3360",
    "basePath": "/",
    "paths": {
...
        "/api/auth/verify-email/send": {
            "post": {
                "description": "Sends a verification email to the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Send Verification Email",
                "parameters": [
                    {
                        "description": "Send Verification Email Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SendVerificationEmailData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Send Verification Email Result",
                        "schema": {
                            "$ref": "#/definitions/handlers.SendVerificationEmailResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid Input or User does not exist",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/verify-email/verify": {
            "post": {
                "description": "Verifies the email using the provided token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify Email Token",
                "parameters": [
                    {
                        "description": "Verify Email Token Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyEmailTokenData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verify Email Token Result",
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyEmailTokenResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid Input or Token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.SendVerificationEmailData": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "handlers.SendVerificationEmailResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.VerifyEmailTokenData": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "handlers.VerifyEmailTokenResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "debug": {
                    "description": "DebugMessage is a technical message that can be used for debugging.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a user-friendly message that can be displayed to the end user.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-API-Key",
            "in": "header"
        },
        "AdminHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Admin-Token",
            "in": "header"
        },
        "RefreshHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Refresh-Token",
            "in": "header"
        },
        "SessionHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Session-Token",
            "in": "header"
        }
    },
...
}